[42]: move uicomponents
