+++ app/src/main/java/com/github/tinkzhang/readkeeper/reading/ReadingListCard.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavController
 import com.github.tinkzhang.readkeeper.model.ReadingBook
 import com.github.tinkzhang.readkeeper.model.ReadingBookFactory
-import com.github.tinkzhang.readkeeper.reading.uicomponents.BookListCard
+import com.github.tinkzhang.uicomponent.BookListCard
 import com.github.tinkzhang.readkeeper.reading.uicomponents.ReadingCardEditBottom
 import com.github.tinkzhang.readkeeper.reading.uicomponents.ReadingCardMetadata
 import com.github.tinkzhang.readkeeper.reading.uicomponents.ReadingCardProgressBottom
