--- app/src/main/java/com/github/tinkzhang/readkeeper/reading/uicomponents/BookCardImage.kt
@@ -1,23 +1,0 @@
-package com.github.tinkzhang.readkeeper.reading.uicomponents
-
-//@Composable
-//fun BookCardImage(book: Book, modifier: Modifier = Modifier) {
-//    Image(
-//        painter = rememberImagePainter(
-//            data = book.bookInfo.imageUrl,
-//            builder = {
-//                this.crossfade(true)
-//                placeholder(drawableResId = R.drawable.ic_launcher_foreground)
-//            }
-//        ),
-//        contentDescription = book.bookInfo.title,
-//        contentScale = ContentScale.Fit,
-//        modifier = modifier.size(128.dp)
-//    )
-//}
-//
-//@Preview
-//@Composable
-//private fun BookCardImagePreview() {
-//    BookCardImage(book = ReadingBookFactory.buildSample())
-//}
