+++ app/src/main/java/com/github/tinkzhang/readkeeper/reading/uicomponents/ReadingCardProgressBottom.kt
@@ -15,6 +15,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.tinkzhang.readkeeper.model.ReadingBook
 import com.github.tinkzhang.readkeeper.model.ReadingBookFactory
+import com.github.tinkzhang.uicomponent.ReadingProgressBar
+import com.github.tinkzhang.uicomponent.ReadingProgressText
 
 @Composable
 fun ReadingCardProgressBottom(
