+++ app/src/main/java/com/github/tinkzhang/readkeeper/search/SearchListCard.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.tinkzhang.readkeeper.model.SearchBook
 import com.github.tinkzhang.readkeeper.model.SearchBookFactory
-import com.github.tinkzhang.readkeeper.reading.uicomponents.BookListCard
+import com.github.tinkzhang.uicomponent.BookListCard
 import com.github.tinkzhang.readkeeper.search.components.SearchCardBottom
 import com.github.tinkzhang.readkeeper.search.components.SearchCardMetadata
 import com.github.tinkzhang.uicomponent.BookCardImage
