+++ common/uicomponent/src/main/java/com/github/tinkzhang/uicomponent/ReadingProgressBar.kt
@@ -1,4 +1,4 @@
-package com.github.tinkzhang.readkeeper.reading.uicomponents
+package com.github.tinkzhang.uicomponent
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.height
